TASK-create-service-layer
Added task state service and refactored task state controller
